// proj_monitoramento/backend/src/modules/user/Interfaces/ICreateUserAlimentadorDTO.ts
interface ICreateUserAlimentadorDTO {
 
  nome_alimentador: string; 
  nome_orgao: string;
  cargo: string;
  profissao: string;
  nivel_escolaridade: string;
  sexo: string;
  genero: string;
  cpf_alimentador: string; 
  mat_orgao: string;
  unidade_sei: string;
  email_institucional: string;
  email_pessoal: string;
  cnpj_alimentador: string;
  nome_empresa: string;
  num_contrato: string;
  password: string;
  

  }
  
  export { ICreateUserAlimentadorDTO };
  